+++ config/routes.rb
@@ -9,7 +9,8 @@ Rails.application.routes.draw do
 
   resources :survivors, only: %i[create update]
 
-  post '/survivors/report', to: 'survivors#report'
+  post '/survivors/:id/trade', to: 'survivors#trade'
+  post '/survivors/:id/report', to: 'survivors#report'
 
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end